Updated IMLBuildManager to copy iml files to persistent location on Android

+++ Assets/InteractML/Scripts/BuildSettings/Editor/IMLBuildManager.cs
@@ -13,8 +13,10 @@ public class IMLBuildManager
     [InitializeOnLoadMethod]
     private static void Initialize()
     {
+#if UNITY_STANDALONE || UNITY_ANDROID 
         // Subscribe to event that fires when a build is produced
         BuildPlayerWindow.RegisterBuildPlayerHandler(BuildPlayerHandler);
+#endif
     }
 
     /// <summary>
@@ -26,15 +28,21 @@ public class IMLBuildManager
         // Build the player as set by the unity buildplayer window
         BuildPlayerWindow.DefaultBuildMethods.BuildPlayer(options);
 
+        // Calculate the datapath for the build
+        string buildDataPath = "";
+#if UNITY_STANDALONE
         // Get filename.
         string buildPath = options.locationPathName;        
         // Cut name of executable from path
         string executableName = Path.GetFileNameWithoutExtension(buildPath);
         buildPath = Path.GetDirectoryName(buildPath);
-        string buildDataPath = Path.Combine(buildPath, executableName + "_Data");
+        buildDataPath = Path.Combine(buildPath, executableName + "_Data");
         //UnityEngine.Debug.Log("=========CUSTOM BUILD CODE==========");
         //UnityEngine.Debug.Log("The app will be built at: " + buildPath);
-
+#elif UNITY_ANDROID
+        // The datapath for the build will be the persistent one
+        buildDataPath = Application.persistentDataPath;
+# endif
         // Copy models and training data from the project folder to the build folder, alongside the built game.
         // Create IML directory
         if (!Directory.Exists(buildDataPath + "/InteractML/Data"))
@@ -47,36 +55,46 @@ public class IMLBuildManager
         // Calculate target paths
         string targetPathModels = buildDataPath + "/InteractML/Data/Models";
         string targetPathTrainingExamples = buildDataPath + "/InteractML/Data/Training_Examples";
-        // Get all modelPaths
-        string[] modelNames = Directory.GetFiles("Assets/InteractML/Data/Models");
-        // Copy the models and overwrite destination files if they already exist.
-        foreach (string modelName in modelNames)
+        
+        // If the Models folder exists in current project...
+        if (Directory.Exists("Assets/InteractML/Data/Models"))
         {
-            // Use static Path methods to extract only the file name from the path.
-            string fileName = Path.GetFileName(modelName);
-            string destFile = Path.Combine(targetPathModels, fileName);
-            File.Copy(modelName, destFile, true);
-        }
-        // Get all trainingExamplesPaths
-        string[] trainingExamplesNames = Directory.GetFiles("Assets/InteractML/Data/Training_Examples");
-        // Copy the models and overwrite destination files if they already exist.
-        foreach (string tExamplesName in trainingExamplesNames)
-        {
-            // Use static Path methods to extract only the file name from the path.
-            string fileName = Path.GetFileName(tExamplesName);
-            string destFile = Path.Combine(targetPathTrainingExamples, fileName);
-            try
-            {               
-                File.Copy(tExamplesName, destFile, true);
-            }
-            catch (System.Exception e)
+            // Get all modelPaths
+            string[] modelNames = Directory.GetFiles("Assets/InteractML/Data/Models");
+            // Copy the models and overwrite destination files if they already exist.
+            foreach (string modelName in modelNames)
             {
-
-                UnityEngine.Debug.LogError(e.Message);
-                
+                // Use static Path methods to extract only the file name from the path.
+                string fileName = Path.GetFileName(modelName);
+                string destFile = Path.Combine(targetPathModels, fileName);
+                File.Copy(modelName, destFile, true);
             }
-            
         }
+
+        // If the Training Examples folder exists in current project...
+        if (Directory.Exists("Assets/InteractML/Data/Training_Examples"))
+        {
+            // Get all trainingExamplesPaths
+            string[] trainingExamplesNames = Directory.GetFiles("Assets/InteractML/Data/Training_Examples");
+            // Copy the models and overwrite destination files if they already exist.
+            foreach (string tExamplesName in trainingExamplesNames)
+            {
+                // Use static Path methods to extract only the file name from the path.
+                string fileName = Path.GetFileName(tExamplesName);
+                string destFile = Path.Combine(targetPathTrainingExamples, fileName);
+                try
+                {
+                    File.Copy(tExamplesName, destFile, true);
+                }
+                catch (System.Exception e)
+                {
+
+                    UnityEngine.Debug.LogError(e.Message);
+
+                }
+
+            }
+        }        
 
         //UnityEngine.Debug.Log("=========END OF CUSTOM BUILD CODE==========");
